+++ QiniuSDK/Common/QNALAssetFile.h
@@ -10,6 +10,7 @@
 
 #import "QNFileDelegate.h"
 
+#if defined(__IPHONE_OS_VERSION_MAX_ALLOWED)
 @class ALAsset;
 @interface QNALAssetFile : NSObject <QNFileDelegate>
 
@@ -24,3 +25,4 @@
 - (instancetype)init:(ALAsset *)asset
                error:(NSError *__autoreleasing *)error;
 @end
+#endif